Fix rubber banding in grid mode with stricter matching.
Turns out the rubber banding in grid mode wasn't working as expected as the early vs. late video count was off by a factor of two. This was fixed with stricter matching that now also takes the class into account.
Futhermore the code got refactored so that the rubber banding and matching is easier to understand. Last but not least the code now logs a warning in case the early/late video counts don't match once all videos have been loaded..